requirements.txt
alembic==1.8.1
attrs==22.1.0
autopage==0.5.1
Boruta==0.3
hyperopt==0.2.7
imbalanced-learn==0.9.1
imblearn==0.0
matplotlib==3.5.3
numba==0.56.0
numpy==1.21.4
optuna==3.0.0
pyaml==21.10.1
PyYAML==6.0
scikit-learn
pandas
shap==0.41.0
xgboost==1.6.2
lightgbm
seaborn==0.12.0
lightgbm==3.3.2
category-encoders==2.5.0